TST: Split tests into smaller files, use more fixtures
